+++ CHANGELOG.md
@@ -22,7 +22,11 @@ For more information on SemVer, please visit http://semver.org/.
 
 
 ## Version 1.4.15
-### 2014-01-??
+### 2014-01-31
+
+#### Bugs
+
+* Fixed static declaration
 
 #### Experimental